Keep UFF evaluation precision-native

+++ src/forcefields/uff_kernels.h
@@ -257,7 +257,7 @@ cudaError_t launchBlockPerMolEnergyKernel(int                                   
                                           const EnergyForceContribsDevicePtrSingle& terms,
                                           const BatchedIndicesDevicePtr&            sytemIndices,
                                           const float*                              coords,
-                                          double*                                   energies,
+                                          float*                                    energies,
                                           bool                                      hasConstraints,
                                           cudaStream_t                              stream           = nullptr,
                                           const uint8_t*                            activeSystemMask = nullptr);
